+++ R/tbl_svysummary.R
@@ -117,7 +117,7 @@ tbl_svysummary <- function(data,
     )
   )
 
-  data <- .svy_ignore_missing_by_obs(data, by = by)
+  data <- .svy_ignore_missing_by_obs(data, by = by, include)
   include <- setdiff(include, by) # remove by variable from list vars included
 
   if (missing(missing)) {
@@ -383,7 +383,7 @@ tbl_svysummary <- function(data,
   x
 }
 
-.svy_ignore_missing_by_obs <- function(data, by) {
+.svy_ignore_missing_by_obs <- function(data, by, include) {
   if (is_empty(by) || !any(is.na(data$variables[[by]]))) {
     return(data)
   }
@@ -393,7 +393,20 @@ tbl_svysummary <- function(data,
     "{.val {sum(obs_to_drop)}} row{?s} with missingness in the {.val {by}} column
     {cli::qty(sum(obs_to_drop))}{?has/have} been removed with {.fun subset}."
   )
-  call2("subset", x = expr(data), subset = expr(!is.na(!!sym(by)))) |>
-    eval()
-}
 
+  # save original labels (subsetting removes labels)
+  original_lbs <- lapply(data$variables[c(by, include)], \(x) {attr(x, "label")})
+
+  # subset data
+  data <-
+    call2("subset", x = expr(data), subset = expr(!is.na(!!sym(by)))) |>
+    eval()
+
+  # restore column labels
+  for (v in c(by, include)) {
+    attr(data$variables[[v]], "label") <- original_lbs[[v]]
+  }
+
+  # return data
+  data
+}