+++ tests/testthat/test-tbl_svysummary.R
@@ -628,10 +628,9 @@ test_that("tbl_svysummary() missing `by` handling", {
   database <- data.frame(
     INDIV_AGE = rnorm(100, mean = 50, sd = 4),
     INDIV_GENDER = rbinom(n = 100, size=1, prob = 0.6),
-    PAIN_SCALE = factor(sample(c("Low", "Elevated"), size = 100, replace = T)),
-    FLOWER_COLOR = factor(sample(c("Blue", "Red"), size = 100, replace = T)),
     poids = rnorm(100, mean = 2, sd = 0.8)
   )
+  attr(database$INDIV_AGE, "label") <- "Subject Age"
 
   # no errors and proper messaging
   expect_snapshot(
@@ -656,5 +655,11 @@ test_that("tbl_svysummary() missing `by` handling", {
     ) |>
       tbl_svysummary(by = "INDIV_GENDER") |>
       as.data.frame()
+  )
+
+  # check the variable label has been retained
+  expect_equal(
+    tbl$table_body$label[1],
+    "Subject Age"
   )
 })