+++ NEWS.md
@@ -1,5 +1,7 @@
 # gtsummary (development version)
 
+* Fix in `tbl_svysummary()` when there are missing values in the `by` column. 
+
 * Added the `"add_overall.tbl_summary-arg:col_label"` theme element to control the default value of `add_overall.tbl_summary(col_label)`. (#2246)
 
 * Added `tbl_split_by_rows()` and `tbl_split_by_columns()` to split tables horizontally (row-wise) and vertically (column-wise). (#2216)